% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data2mark}
\alias{data2mark}
\title{Prepares data to be used with MARK software through RMark, uses calendar week as sampling instances}
\usage{
data2mark(data, model)
}
\arguments{
\item{data}{Data with rows corresponding to observations and columns id, date, sex}

\item{model}{RMark model type}
}
\value{
A list suitable for use with RMark
}
\description{
Prepares data to be used with MARK software through RMark, uses calendar week as sampling instances
}
