% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_mark}
\alias{fit_mark}
\title{Fits a range of models using RMark}
\usage{
fit_mark(data)
}
\arguments{
\item{data}{Data with rows corresponding to observations and columns id, date, sex}
}
\value{
A table of fitted models and summaries
}
\description{
Fits a range of models using RMark
}
