% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forward}
\alias{forward}
\title{Simulates a population forwards in time}
\usage{
forward(model_fits, rate, hunted)
}
\arguments{
\item{rate}{Natural growth rate}

\item{hunted}{Yearly and regional hunting data}

\item{model_fit}{Fitted models}
}
\value{
A table simulated population sizes
}
\description{
Simulates a population forwards in time
}
