% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulates from a fitted object}
\usage{
sim_pop(fit)
}
\arguments{
\item{fit}{Fitted models}
}
\value{
A table simulated population sizes
}
\description{
Simulates from a fitted object
}
